move transactionFilter to fetchcontracts
